Added fire button functionality
the fire button now draws the laser line of the solution, does not draw incorrect solution lines.. yet. Might want to consider adding a timer so the line clears it self after a few seconds.. and adding in the lines that could be drawn if the incorrect orientation of mirrors is made. Also some type of dialog box that tells the user they are correct or incorrect.