Intermediate Board Panel - Incomplete
This is the next level of the board panel, has rotational functionality and drag and drop functionality for one of the mirrors. Still needed are drag and drop of the other two mirrors and to draw the correct laser path when the fire button is pressed. Included are the two green mirror images needed. my images are kept in the same directory as the class so if you keep your images inside a folder you will need to add "Images\\" in front of all the file paths for the images. -Patrick